requirements.txt
pip
pyyaml
telethon
cryptg
Tgcrypto
pillow
aiohttp
apscheduler
hachoir
loguru
envparse
validators
redis
aioredis
requests
redis
telegraph
teleredis
markdown
ffmpeg
motor
pymongo[srv]
asyncio
names
random-address
fuzzywuzzy
# python-Levenshtein
opencv-python
qrcode
pillow
bs4
aiofiles
pytesseract
aiofiles
numpy
pytesseract
qrcode
googletrans
selenium
lxml
pytube
numerize 
gTTS
faker
emoji
fontTools
git+https://github.com/66niko99/py-adyen-encrypt.git
pymongo